@@ -1,26 +1,19 @@
-buildscript {
-    repositories {
-        google()
-    }
-}
-
 apply plugin: 'com.android.application'
 apply plugin: 'com.google.firebase.crashlytics'
 apply plugin: 'com.google.gms.google-services'
 
 repositories {
-    maven { url 'https://maven.google.com' }
     maven { url 'https://jitpack.io' }
 }
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion '29.0.0'
+    compileSdkVersion 30
+    buildToolsVersion '30.0.2'
 
     defaultConfig {
         applicationId project.property('demo_applicationId')
         minSdkVersion 16
-        targetSdkVersion 29
+        targetSdkVersion 30
         versionCode project.property('versionCode').toString().toInteger()
         versionName project.property('versionName')
         vectorDrawables {
@@ -54,6 +47,10 @@ android {
     lintOptions {
         abortOnError false
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
@@ -72,4 +69,5 @@ dependencies {
     implementation 'androidx.recyclerview:recyclerview:1.1.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.preference:preference:1.1.1'
 }
