@@ -6,18 +6,14 @@ import androidx.annotation.Nullable;
 import com.webimapp.android.sdk.Message;
 import com.webimapp.android.sdk.MessageListener;
 import com.webimapp.android.sdk.MessageTracker;
-import com.webimapp.android.sdk.Webim;
-import com.webimapp.android.sdk.impl.backend.WebimInternalLog;
 import com.webimapp.android.sdk.impl.items.ChatItem;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
@@ -28,11 +24,12 @@ public class MessageHolderImpl implements MessageHolder {
     private final RemoteHistoryProvider historyProvider;
 
     private final HistoryStorage historyStorage;
-    private final List<MessageImpl> currentChatMessages = new ArrayList<>();
+    private final List<MessageImpl> currentMessages = new ArrayList<>();
     private final List<MessageSending> sendingMessages = new ArrayList<>();
     private boolean isReachedEndOfLocalHistory = false;
     private boolean isReachedEndOfRemoteHistory;
-    private int lastChatMessageIndex = 0;
+    private boolean isFirstUpdateReceived;
+    private int previousChatLastMessageIndex = 0;
     @Nullable
     private MessageTrackerImpl messageTracker;
 
@@ -54,77 +51,40 @@ public class MessageHolderImpl implements MessageHolder {
     public void receiveHistoryUpdate(List<? extends MessageImpl> messages,
                                      Set<String> deleted,
                                      final Runnable callback) {
-        historyStorage.receiveHistoryUpdate(messages, deleted,
-                new HistoryStorage.UpdateHistoryCallback() {
-                    @Override
-                    public void onHistoryChanged(@NonNull MessageImpl message) {
-                        if (messageTracker != null) {
-                            messageTracker.onHistoryChanged(message);
-                        }
+        historyStorage.receiveHistoryUpdate(
+            messages,
+            deleted,
+            new HistoryStorage.UpdateHistoryCallback() {
+                @Override
+                public void onHistoryAdded(@Nullable String beforeId, @NonNull MessageImpl historyMessage) {
+                    if (messageTracker != null) {
+                        messageTracker.onHistoryAdded(beforeId, historyMessage);
+                    }
+                }
+
+                @Override
+                public void onHistoryChanged(@NonNull MessageImpl historyMessage) {
+                    if (messageTracker != null) {
+                        messageTracker.onHistoryChanged(historyMessage);
+                    }
+                }
+
+                @Override
+                public void onHistoryDeleted(String serverSideId) {
+                    if (messageTracker != null) {
+                        messageTracker.onHistoryDeleted(serverSideId);
+                    }
+                }
+
+                @Override
+                public void endOfBatch() {
+                    if (messageTracker != null) {
+                        messageTracker.onHistoryEndOfBatch();
                     }
 
-                    @Override
-                    public void onHistoryAdded(@Nullable HistoryId before, @NonNull MessageImpl message) {
-                        if (!tryMergeWithLastChat(message) && (messageTracker != null)) {
-                            messageTracker.onHistoryAdded(before, message);
-                        }
-                    }
-
-                    private boolean tryMergeWithLastChat(@NonNull MessageImpl historyMessage) {
-                        for (int i = 0; i < currentChatMessages.size(); i++) {
-                            MessageImpl chatMessage = currentChatMessages.get(i);
-                            if (chatMessage.getId().equals(historyMessage.getId())) {
-                                if (i < lastChatMessageIndex) {
-                                    MessageImpl replacementMessage
-                                            = chatMessage.transferToHistory(historyMessage);
-                                    currentChatMessages.remove(i);
-                                    lastChatMessageIndex--;
-
-                                    if (messageTracker != null) {
-                                        messageTracker.idToHistoryMessageMap.put(
-                                                historyMessage.getHistoryId().getDbId(),
-                                                replacementMessage
-                                        );
-
-                                        if (replacementMessage != chatMessage) {
-                                            messageTracker.messageListener.messageChanged(
-                                                    chatMessage,
-                                                    replacementMessage
-                                            );
-                                        }
-                                    }
-                                } else {
-                                    chatMessage.setSecondaryHistory(historyMessage);
-                                    if (messageTracker != null) {
-                                        messageTracker.idToHistoryMessageMap.put(
-                                                historyMessage.getHistoryId().getDbId(),
-                                                historyMessage
-                                        );
-                                    }
-                                }
-
-                                return true;
-                            }
-                        }
-                        return false;
-                    }
-
-                    @Override
-                    public void onHistoryDeleted(String id) {
-                        if (messageTracker != null) {
-                            messageTracker.onHistoryDeleted(id);
-                        }
-                    }
-
-                    @Override
-                    public void endOfBatch() {
-                        if (messageTracker != null) {
-                            messageTracker.onHistoryEndOfBatch();
-                        }
-
-                        callback.run();
-                    }
-                });
+                    callback.run();
+                }
+            });
     }
 
     public void setReachedEndOfRemoteHistory(boolean reachedEndOfRemoteHistory) {
@@ -142,8 +102,8 @@ public class MessageHolderImpl implements MessageHolder {
     }
 
     private void getLatestMessages(int limit, @NonNull MessageTracker.GetMessagesCallback callback) {
-        if (currentChatMessages.size() != 0) {
-            respondMessages(callback, currentChatMessages, limit);
+        if (!currentMessages.isEmpty()) {
+            respondMessages(callback, currentMessages, limit);
         } else {
             historyStorage.getLatest(limit, callback);
         }
@@ -152,53 +112,43 @@ public class MessageHolderImpl implements MessageHolder {
     private void getMessages(@NonNull MessageImpl before,
                              int limit,
                              @NonNull final MessageTracker.GetMessagesCallback callback) {
-        if (before.getSource().isCurrentChat()) {
-            if (currentChatMessages.size() == 0) {
-                throw new RuntimeException("Requested history related to " +
-                        "@CurrentChat AbstractMessage when current chat is empty");
-            }
-            @CurrentChat MessageImpl first = currentChatMessages.get(0);
-            if (before == first) {
-                if (!first.hasHistoryComponent()) {
-                    historyStorage.getLatest(limit, callback);
-                } else {
-                    getMessagesFromHistory(first.getHistoryId(), limit, callback);
-                }
+        if (!currentMessages.isEmpty()) {
+            @CurrentChat MessageImpl first = currentMessages.get(0);
+            if (before.equals(first)) {
+                getMessagesFromHistory(first, limit, callback);
             } else {
-                getMessagesFromCurrentChat(before, limit, callback);
+                getMessagesFromCurrent(before, limit, callback);
             }
         } else {
-            getMessagesFromHistory(before.getHistoryId(), limit, callback);
+            getMessagesFromHistory(before, limit, callback);
         }
     }
 
-    private void getMessagesFromCurrentChat(@NonNull @CurrentChat MessageImpl before,
-                                            int limit,
-                                            @NonNull final MessageTracker.GetMessagesCallback
+    private void getMessagesFromCurrent(@NonNull @CurrentChat MessageImpl before,
+                                        int limit,
+                                        @NonNull final MessageTracker.GetMessagesCallback
                                                     callback) {
-        before.getSource().assertCurrentChat();
-        int ind = currentChatMessages.indexOf(before);
+        int ind = currentMessages.indexOf(before);
         if (ind <= 0) {
             String sessionId = before.sessionId;
-            String searchMessageId = before.getId().toString();
+            String searchMessageId = before.getClientSideId().toString();
             String borderChatMessagesId =
-                    currentChatMessages.get(0).getId().toString() +
+                    currentMessages.get(0).getClientSideId().toString() +
                     " / " +
-                    currentChatMessages.get(currentChatMessages.size() - 1).getId().toString();
+                    currentMessages.get(currentMessages.size() - 1).getClientSideId().toString();
             throw new IllegalStateException(
                     "Impossible: " +
                     sessionId + " / " +
                     searchMessageId + " / " +
                     borderChatMessagesId);
         }
-        respondMessages(callback, currentChatMessages, ind, limit);
+        respondMessages(callback, currentMessages, ind, limit);
     }
 
-    private void getMessagesFromHistory(@NonNull final HistoryId before,
+    private void getMessagesFromHistory(@NonNull final MessageImpl before,
                                         final int limit,
-                                        @NonNull final MessageTracker.GetMessagesCallback
-                                                callback) {
-        if (!isReachedEndOfLocalHistory) {
+                                        @NonNull final MessageTracker.GetMessagesCallback callback) {
+        if (before.isSavedInHistory() && !isReachedEndOfLocalHistory) {
             historyStorage.getBefore(before, limit, new MessageTracker.GetMessagesCallback() {
                 @Override
                 public void receive(@NonNull final List<? extends Message> messages) {
@@ -206,44 +156,47 @@ public class MessageHolderImpl implements MessageHolder {
                         isReachedEndOfLocalHistory = true;
                         getMessagesFromHistory(before, limit, callback);
                     } else if (messages.size() < limit && !isReachedEndOfRemoteHistory) {
-                        final List<Message> result =
-                                Collections.synchronizedList(new ArrayList<>(messages));
-                        requestHistoryBefore(((MessageImpl) messages.get(0)).getHistoryId(),
-                                limit - messages.size(),
-                                new MessageTracker.GetMessagesCallback() {
-                                    @Override
-                                    public void receive(@NonNull List<? extends Message> rest) {
-                                        result.addAll(rest);
-                                        callback.receive(unmodifiableList(result));
-                                    }
-                                });
+                        requestMessagesFromRemoteHistory(messages);
                     } else {
                         callback.receive(messages);
                     }
                 }
+
+                private void requestMessagesFromRemoteHistory(@NonNull List<? extends Message> messages) {
+                    final List<Message> result =
+                        Collections.synchronizedList(new ArrayList<>(messages));
+                    requestHistoryBefore(
+                        ((MessageImpl) messages.get(0)),
+                        limit - messages.size(),
+                        remoteMessages -> {
+                            result.addAll(remoteMessages);
+                            callback.receive(unmodifiableList(result));
+                        }
+                    );
+                }
             });
         } else if (isReachedEndOfRemoteHistory) {
-            callback.receive(Collections.<Message>emptyList());
+            callback.receive(Collections.emptyList());
         } else {
             requestHistoryBefore(before, limit, callback);
         }
     }
 
-    private void requestHistoryBefore(@NonNull HistoryId before,
+    private void requestHistoryBefore(@NonNull MessageImpl before,
                                       final int limit,
                                       @NonNull final MessageTracker.GetMessagesCallback callback) {
-        historyProvider.requestHistoryBefore(before.getTimeMicros(), new HistoryBeforeCallback() {
-            @Override
-            public void onSuccess(List<? extends MessageImpl> messages, boolean hasMore) {
+        historyProvider.requestHistoryBefore(
+            before.getTimeMicros(),
+            (messages, hasMore) -> {
                 if (messages.isEmpty()) {
                     isReachedEndOfRemoteHistory = true;
                 } else {
-                    isReachedEndOfLocalHistory = false;
+                    isReachedEndOfRemoteHistory = false;
                     historyStorage.receiveHistoryBefore(messages, hasMore);
                 }
                 respondMessages(callback, messages, limit);
             }
-        });
+        );
     }
 
     private static void respondMessages(MessageTracker.GetMessagesCallback callback,
@@ -274,12 +227,12 @@ public class MessageHolderImpl implements MessageHolder {
     public void onChatReceive(@Nullable ChatItem oldChat,
                               @Nullable ChatItem newChat,
                               @CurrentChat List<? extends MessageImpl> newMessages) {
-        if (currentChatMessages.isEmpty()) {
+        if (currentMessages.isEmpty() || (oldChat == null && newChat != null)) {
             receiveNewMessages(newMessages);
         } else {
             if (newChat == null) {
                 historifyCurrentChat();
-            } else if (oldChat == null || !InternalUtils.equals(oldChat, newChat)) {
+            } else if (!InternalUtils.equals(oldChat, newChat)) {
                 historifyCurrentChat();
                 receiveNewMessages(newMessages);
             } else {
@@ -289,124 +242,112 @@ public class MessageHolderImpl implements MessageHolder {
     }
 
     private void historifyCurrentChat() {
-        for (Iterator<MessageImpl> iterator = currentChatMessages.iterator(); iterator.hasNext(); ) {
-            MessageImpl chatMessage = iterator.next();
-            if (chatMessage.hasHistoryComponent()) {
-                chatMessage.invert();
-                if (messageTracker != null) {
-                    String id = chatMessage.getHistoryId().getDbId();
-                    MessageImpl historyMessage = messageTracker.idToHistoryMessageMap.get(id);
-                    if ((historyMessage != null)) {
-                        historyMessage.setCanBeEdited(false);
-                        historyMessage.setCanBeReplied(false);
-                        if (!MessageImpl.isContentEquals(chatMessage, historyMessage)) {
-                            messageTracker.messageListener.messageChanged(chatMessage, historyMessage);
-                        } else {
-                            messageTracker.idToHistoryMessageMap.put(id, chatMessage);
-                        }
-                    }
-                }
-                iterator.remove();
-            } else {
-                if (messageTracker != null && (chatMessage.canBeEdited() || chatMessage.canBeReplied())) {
-                    chatMessage.setCanBeEdited(false);
-                    chatMessage.setCanBeReplied(false);
-                    messageTracker.messageListener.messageChanged(chatMessage, chatMessage);
-                }
+        for (MessageImpl chatMessage : currentMessages) {
+            if (messageTracker != null && (chatMessage.canBeEdited() || chatMessage.canBeReplied())) {
+                chatMessage.setCanBeEdited(false);
+                messageTracker.onCurrentMessageChanged(chatMessage, chatMessage);
             }
         }
 
-        lastChatMessageIndex = currentChatMessages.size();
+        previousChatLastMessageIndex = currentMessages.size();
     }
 
-    private void mergeCurrentChatWith(@CurrentChat List<? extends MessageImpl> newMessages) {
-        int oldMessageIndex = this.lastChatMessageIndex;
-        boolean isOldMessagesEnded = false;
-        for (int newMessageIndex = 0; (newMessageIndex < newMessages.size()); newMessageIndex++) {
-            MessageImpl newMessage = newMessages.get(newMessageIndex);
-            if (!isOldMessagesEnded) {
-                boolean merged = false;
-                while (oldMessageIndex < currentChatMessages.size()) {
-                    MessageImpl oldMessage = currentChatMessages.get(oldMessageIndex);
-                    if (oldMessage.getId().equals(newMessage.getId())) {
-                        if (!MessageImpl.isContentEquals(oldMessage, newMessage)) {
-                            currentChatMessages.set(oldMessageIndex, newMessage);
-                            if (messageTracker != null) {
-                                messageTracker.onCurrentChatMessageChanged(
-                                        oldMessageIndex,
-                                        oldMessage,
-                                        newMessage
-                                );
-                            }
-                        }
-                        merged = true;
-                        oldMessageIndex++;
+    private void mergeCurrentChatWith(@CurrentChat List<? extends MessageImpl> newChatMessages) {
+        int oldMessageIndex = this.previousChatLastMessageIndex;
+        List<MessageImpl> messagesToDelete = new ArrayList<>(currentMessages.subList(oldMessageIndex, currentMessages.size()));
 
-                        break;
-                    } else {
-                        currentChatMessages.remove(oldMessageIndex);
-
+        for (MessageImpl newMessage : newChatMessages) {
+            boolean merged = false;
+            while (oldMessageIndex < currentMessages.size()) {
+                MessageImpl currentMessage = currentMessages.get(oldMessageIndex);
+                if (currentMessage.equals(newMessage)) {
+                    // update currentMessage
+                    if (!currentMessage.isContentEquals(newMessage)) {
+                        currentMessages.set(oldMessageIndex, newMessage);
                         if (messageTracker != null) {
-                            messageTracker.onCurrentChatMessageDeleted(oldMessageIndex, oldMessage);
+                            messageTracker.onCurrentMessageChanged(currentMessage, newMessage);
                         }
                     }
+                    merged = true;
+                    messagesToDelete.remove(currentMessage);
+                    oldMessageIndex++;
+                    break;
+                } else if (currentMessage.getTimeMicros() >= newMessage.getTimeMicros()) {
+                    break;
                 }
-                if (!merged && (oldMessageIndex >= currentChatMessages.size())) {
-                    isOldMessagesEnded = true;
+                oldMessageIndex++;
+            }
+
+            if (!merged) {
+                int nextMessageIndex = oldMessageIndex + 1;
+                MessageImpl beforeMessage = getNextCurrentMessage(nextMessageIndex);
+                currentMessages.add(oldMessageIndex, newMessage);
+                if (messageTracker != null) {
+                    messageTracker.onCurrentMessageAdded(beforeMessage, newMessage);
+                }
+                oldMessageIndex++;
+            }
+        }
+
+        if (!messagesToDelete.isEmpty()) {
+            for (int i = 0; i < messagesToDelete.size(); i++) {
+                MessageImpl needDeleteMessage = messagesToDelete.get(i);
+                currentMessages.remove(needDeleteMessage);
+                if (messageTracker != null) {
+                    messageTracker.onCurrentMessageDeleted(i, needDeleteMessage);
                 }
             }
-            if (isOldMessagesEnded) {
-                receiveNewMessage(newMessage);
+        }
+    }
+
+    @Nullable
+    private MessageImpl getNextCurrentMessage(int nextMessageIndex) {
+        return currentMessages.size() > nextMessageIndex ? currentMessages.get(nextMessageIndex) : null;
+    }
+
+    private void receiveNewMessages(@NonNull @CurrentChat List<? extends MessageImpl> newMessages) {
+        if (!newMessages.isEmpty()) {
+            isFirstUpdateReceived = true;
+            if (messageTracker != null) {
+                messageTracker.onNewMessages(newMessages);
+            } else {
+                currentMessages.addAll(newMessages);
             }
         }
     }
 
     @Override
-    public void receiveNewMessage(@NonNull @CurrentChat MessageImpl msg) {
+    public void onMessageAdded(@NonNull @CurrentChat MessageImpl msg) {
         if (messageTracker != null) {
             messageTracker.onNewMessage(msg);
-        }
-        else {
-            currentChatMessages.add(msg);
-        }
-    }
-
-    private void receiveNewMessages(@NonNull @CurrentChat List<? extends MessageImpl> newMessages) {
-        if (!newMessages.isEmpty()) {
-            if (messageTracker != null) {
-                messageTracker.onNewMessages(newMessages);
-            } else {
-                currentChatMessages.addAll(newMessages);
-            }
+        } else {
+            currentMessages.add(msg);
         }
     }
 
     @Override
     public void onMessageChanged(@NonNull @CurrentChat MessageImpl newMessage) {
-        for (int i = lastChatMessageIndex; i < currentChatMessages.size(); i++) {
-            @CurrentChat MessageImpl oldMessage = currentChatMessages.get(i);
-            if (oldMessage.getIdInCurrentChat().equals(newMessage.getIdInCurrentChat())) {
-                currentChatMessages.set(i, newMessage);
+        for (int i = previousChatLastMessageIndex; i < currentMessages.size(); i++) {
+            MessageImpl oldMessage = currentMessages.get(i);
+            if (oldMessage.equals(newMessage)) {
+                currentMessages.set(i, newMessage);
                 if (messageTracker != null) {
-                    messageTracker.onCurrentChatMessageChanged(i, oldMessage, newMessage);
+                    messageTracker.onCurrentMessageChanged(oldMessage, newMessage);
                 }
-
                 break;
             }
         }
     }
 
     @Override
-    public void onMessageDeleted(@NonNull String idInCurrentChat) {
-        for (int i = lastChatMessageIndex; i < currentChatMessages.size(); i++) {
-            @CurrentChat MessageImpl oldMessage = currentChatMessages.get(i);
-            if (oldMessage.getIdInCurrentChat().equals(idInCurrentChat)) {
-                currentChatMessages.remove(i);
-
+    public void onMessageDeleted(@NonNull String messageId) {
+        for (int i = previousChatLastMessageIndex; i < currentMessages.size(); i++) {
+            @CurrentChat MessageImpl oldMessage = currentMessages.get(i);
+            if (oldMessage.getServerSideId().equals(messageId)) {
+                currentMessages.remove(i);
                 if (messageTracker != null) {
-                    messageTracker.onCurrentChatMessageDeleted(i, oldMessage);
+                    messageTracker.onCurrentMessageDeleted(i, oldMessage);
                 }
-
                 break;
             }
         }
@@ -415,7 +356,6 @@ public class MessageHolderImpl implements MessageHolder {
     @Override
     public void onSendingMessage(@NonNull MessageSending message) {
         sendingMessages.add(message);
-
         if (messageTracker != null) {
             messageTracker.messageListener.messageAdded(null, message);
         }
@@ -426,8 +366,8 @@ public class MessageHolderImpl implements MessageHolder {
     public String onChangingMessage(@NonNull Message.Id id, @Nullable String text) {
         if (messageTracker != null) {
             MessageImpl message = null;
-            for (MessageImpl curr : currentChatMessages) {
-                if (curr.id.equals(id)) {
+            for (MessageImpl curr : currentMessages) {
+                if (curr.clientSideId.equals(id)) {
                     message = curr;
                     break;
                 }
@@ -438,7 +378,7 @@ public class MessageHolderImpl implements MessageHolder {
 
             MessageImpl newMsg = new MessageImpl(
                     message.serverUrl,
-                    message.id,
+                    message.clientSideId,
                     message.sessionId,
                     message.operatorId,
                     message.avatarUrl,
@@ -446,9 +386,9 @@ public class MessageHolderImpl implements MessageHolder {
                     message.type,
                     text == null ? message.text : text,
                     message.timeMicros,
-                    message.getCurrentChatId(),
+                    message.getServerSideId(),
                     message.getData(),
-                    message.isHistoryMessage,
+                    message.savedInHistory,
                     message.getAttachment(),
                     message.isReadByOperator(),
                     message.canBeEdited(),
@@ -467,14 +407,13 @@ public class MessageHolderImpl implements MessageHolder {
 
     @Override
     public void onMessageSendingCancelled(@NonNull Message.Id id) {
-        for (Iterator<MessageSending> it = sendingMessages.iterator(); it.hasNext(); ) {
-            MessageSending msg = it.next();
-            if (msg.getId().equals(id)) {
-                it.remove();
+        for (Iterator<MessageSending> iterator = sendingMessages.iterator(); iterator.hasNext(); ) {
+            MessageSending msg = iterator.next();
+            if (msg.getClientSideId().equals(id)) {
+                iterator.remove();
                 if (messageTracker != null) {
                     messageTracker.messageListener.messageRemoved(msg);
                 }
-
                 break;
             }
         }
@@ -484,8 +423,8 @@ public class MessageHolderImpl implements MessageHolder {
     public void onMessageChangingCancelled(@NonNull Message.Id id, @NonNull String text) {
         if (messageTracker != null) {
             MessageImpl message = null;
-            for (MessageImpl curr : currentChatMessages) {
-                if (curr.id.equals(id)) {
+            for (MessageImpl curr : currentMessages) {
+                if (curr.clientSideId.equals(id)) {
                     message = curr;
                     break;
                 }
@@ -496,7 +435,7 @@ public class MessageHolderImpl implements MessageHolder {
 
             MessageImpl newMsg = new MessageImpl(
                     message.serverUrl,
-                    message.id,
+                    message.clientSideId,
                     message.sessionId,
                     message.operatorId,
                     message.avatarUrl,
@@ -504,9 +443,9 @@ public class MessageHolderImpl implements MessageHolder {
                     message.type,
                     text,
                     message.timeMicros,
-                    message.getCurrentChatId(),
+                    message.getServerSideId(),
                     message.getData(),
-                    message.isHistoryMessage,
+                    message.savedInHistory,
                     message.getAttachment(),
                     message.isReadByOperator(),
                     message.canBeEdited(),
@@ -528,15 +467,13 @@ public class MessageHolderImpl implements MessageHolder {
 
     @Override
     public boolean historyMessagesEmpty() {
-        return messageTracker == null || messageTracker.idToHistoryMessageMap.isEmpty();
+        return messageTracker == null || currentMessages.isEmpty();
     }
 
     private class MessageTrackerImpl implements MessageTracker {
         private final MessageListener messageListener;
-        private final Map<String, MessageImpl> idToHistoryMessageMap = new HashMap<>();
         private MessageImpl headMessage;
         private boolean isMessagesLoading;
-        private boolean isFirstHistoryUpdateReceived;
         private boolean isAllMessageSourcesEnded;
         private boolean isDestroyed;
 
@@ -565,8 +502,7 @@ public class MessageHolderImpl implements MessageHolder {
                         "Given " + limitOfMessages);
             }
             isMessagesLoading = true;
-            if (isFirstHistoryUpdateReceived ||
-                    (currentChatMessages.size() != 0 && currentChatMessages.get(0) != headMessage)) {
+            if (isFirstUpdateReceived && !currentMessages.isEmpty()) {
                 uncheckedGetNextMessages(limitOfMessages, callback);
             } else {
                 cachedCallback = callback;
@@ -578,15 +514,9 @@ public class MessageHolderImpl implements MessageHolder {
                         if (cachedCallback != null && !messages.isEmpty()) {
                             final GetMessagesCallback callback = cachedCallback;
                             cachedCallback = null;
-                            isFirstHistoryUpdateReceived = true;
-                            new GetMessagesCallbackWrapper(limitOfMessages, new GetMessagesCallback() {
-                                @Override
-                                public void receive(@NonNull List<? extends Message> messages) {
-                                    if (!isDestroyed) {
-                                        callback.receive(messages);
-                                    }
-                                }
-                            }).receive(messages);
+                            cachedLimit = 0;
+                            isFirstUpdateReceived = true;
+                            new GetMessagesCallbackWrapper(limitOfMessages, callback).receive(messages);
                         }
                     }
                 });
@@ -610,7 +540,7 @@ public class MessageHolderImpl implements MessageHolder {
             }
 
             isMessagesLoading = true;
-            currentChatMessages.clear();
+            currentMessages.clear();
             cachedCallback = callback;
             cachedLimit = limitOfMessages;
             isReachedEndOfRemoteHistory = false;
@@ -621,21 +551,16 @@ public class MessageHolderImpl implements MessageHolder {
                 @Override
                 public void receive(@NonNull List<? extends Message> messages) {
                     if (cachedCallback != null
-                            && (!messages.isEmpty() || isFirstHistoryUpdateReceived)) {
+                            && (!messages.isEmpty() || isFirstUpdateReceived)) {
                         final GetMessagesCallback callback = cachedCallback;
                         cachedCallback = null;
-                        isFirstHistoryUpdateReceived = true;
-                        new GetMessagesCallbackWrapper(limitOfMessages, new GetMessagesCallback() {
-                            @Override
-                            public void receive(@NonNull List<? extends Message> messages) {
-                                if (!isDestroyed) {
-                                    callback.receive(messages);
-                                }
-                            }
-                        }).receive(messages);
+                        cachedLimit = 0;
+                        isFirstUpdateReceived = true;
+                        new GetMessagesCallbackWrapper(limitOfMessages, callback).receive(messages);
                     }
                 }
             });
+            currentMessages.clear();
             headMessage = null;
         }
 
@@ -646,13 +571,16 @@ public class MessageHolderImpl implements MessageHolder {
             if (isDestroyed) {
                 throw new IllegalStateException("WebimMessageTracker is destroyed");
             }
+            if (isMessagesLoading) {
+                throw new IllegalStateException("Messages is loading now; " +
+                    "can't load messages in parallel");
+            }
+            isMessagesLoading = true;
 
             historyStorage.getFull(new GetMessagesCallback() {
                 @Override
                 public void receive(@NonNull List<? extends Message> messages) {
-                    if (!isDestroyed) {
-                        callback.receive(messages);
-                    }
+                    new GetMessagesCallbackWrapper(0, callback).receive(messages);
                 }
             });
         }
@@ -684,22 +612,24 @@ public class MessageHolderImpl implements MessageHolder {
             }
 
             MessageImpl unwrappedMessage = (MessageImpl) message;
-            if (unwrappedMessage != headMessage) {
+            if (!unwrappedMessage.equals(headMessage)) {
                 isReachedEndOfLocalHistory = false;
             }
-            if (unwrappedMessage.getSource().isHistory()) {
-                for (Iterator<MessageImpl> iterator = idToHistoryMessageMap.values().iterator();
-                     iterator.hasNext();
-                ) {
-                    MessageImpl iteratedMessage = iterator.next();
-                    if (iteratedMessage.getTimeMicros() < unwrappedMessage.getTimeMicros()) {
-                        iterator.remove();
+            List<MessageImpl> deleted = new ArrayList<>();
+            for (int i = 0; i < currentMessages.size(); i++) {
+                MessageImpl iteratedMessage = currentMessages.get(i);
+                if (iteratedMessage.getTimeMicros() < unwrappedMessage.getTimeMicros()) {
+                    deleted.add(iteratedMessage);
+                    if (previousChatLastMessageIndex >= i) {
+                        previousChatLastMessageIndex--;
                     }
                 }
-            } else {
-                idToHistoryMessageMap.clear();
             }
 
+            for (MessageImpl messageToDelete : deleted) {
+                int index = currentMessages.indexOf(messageToDelete);
+                onCurrentMessageDeleted(index + 1, messageToDelete);
+            }
             headMessage = unwrappedMessage;
         }
 
@@ -727,7 +657,7 @@ public class MessageHolderImpl implements MessageHolder {
                     int sinceMessagePosition = -1;
                     for (int i = 0; i < localStorageMessages.size(); i++) {
                         Message currentLocalMessage = localStorageMessages.get(i);
-                        if (currentLocalMessage.getId().equals(sinceMessage.getId())) {
+                        if (currentLocalMessage.getClientSideId().equals(sinceMessage.getClientSideId())) {
                             sinceMessagePosition = i;
                             break;
                         }
@@ -761,7 +691,7 @@ public class MessageHolderImpl implements MessageHolder {
                                 for (int i = messages.size() - 1; i >= 0; i--) {
                                     MessageImpl historyMessage = messages.get(i);
                                     searchedMessages.add(0, historyMessage);
-                                    if (historyMessage.getId().equals(sinceMessage.getId())) {
+                                    if (historyMessage.getClientSideId().equals(sinceMessage.getClientSideId())) {
                                         headMessage = searchedMessages.get(0);
                                         messagesCallback.receive(searchedMessages);
                                         return;
@@ -783,7 +713,6 @@ public class MessageHolderImpl implements MessageHolder {
         }
 
         void onNewMessage(final @CurrentChat MessageImpl message) {
-            message.getSource().assertCurrentChat();
             addNewOrMergeMessage(message);
             if (headMessage == null && !isAllMessageSourcesEnded) {
                 if (cachedCallback != null) { // TODO on endOfBatch only
@@ -795,175 +724,131 @@ public class MessageHolderImpl implements MessageHolder {
         }
 
         void onNewMessages(final @CurrentChat List<? extends MessageImpl> newMessages) {
-            if (headMessage != null || isAllMessageSourcesEnded) {
+            if (!currentMessages.isEmpty() && headMessage != null) {
                 for (@CurrentChat MessageImpl message : newMessages) {
                     addNewOrMergeMessage(message);
                 }
             } else {
-                currentChatMessages.addAll(newMessages);
+                currentMessages.addAll(newMessages);
                 if (cachedCallback != null) {
                     GetMessagesCallback callback = cachedCallback;
+                    int limit = cachedLimit;
                     cachedCallback = null;
-                    uncheckedGetNextMessages(cachedLimit, callback);
+                    cachedLimit = 0;
+                    uncheckedGetNextMessages(limit, callback);
                 }
             }
         }
 
-        private void addNewOrMergeMessage(final @CurrentChat MessageImpl message) {
-            message.getSource().assertCurrentChat();
-
-            boolean callAdded = true;
-
-            if (headMessage == null) {
-                headMessage = message;
-            } else if (headMessage.getTimeMicros() > message.getTimeMicros()) {
-                callAdded = false;
-
-                currentChatMessages.add(message);
-            } else {
-                // TODO optimize (do no iterate all)
-                for (Iterator<MessageImpl> iterator = idToHistoryMessageMap.values().iterator();
-                     iterator.hasNext(); ) {
-                    MessageImpl historyMessage = iterator.next();
-                    if (message.getId().equals(historyMessage.getId())) {
-                        MessageImpl replacementMessage =
-                                historyMessage.transferToCurrentChat(message);
-                        currentChatMessages.add(replacementMessage);
-
-                        callAdded = false;
-
-                        if (replacementMessage != historyMessage) {
-                            messageListener.messageChanged(historyMessage, replacementMessage);
-                            if (headMessage == historyMessage) {
-                                headMessage = replacementMessage;
-                            }
+        private void addNewOrMergeMessage(final @CurrentChat MessageImpl newMessage) {
+            int currentIndex;
+            boolean addToEnd = true;
+            for (currentIndex = 0; currentIndex < currentMessages.size(); currentIndex++) {
+                MessageImpl message = currentMessages.get(currentIndex);
+                if (newMessage.equals(message)) {
+                    if (!message.isContentEquals(newMessage)) {
+                        currentMessages.set(currentIndex, newMessage);
+                        if (message.isSavedInHistory()) {
+                            newMessage.setSavedInHistory(true);
                         }
-
-                        iterator.remove();
-
-                        break;
+                        onCurrentMessageChanged(message, newMessage);
                     }
+                    return;
+                }
+                if (message.compareTo(newMessage) > 0) {
+                    addToEnd = false;
+                    break;
                 }
             }
 
-            if (callAdded) {
-                currentChatMessages.add(message);
+            MessageSending messageToSend = findMessageSendingMirror(newMessage);
+            if (messageToSend != null) {
+                currentMessages.add(newMessage);
+                messageListener.messageChanged(messageToSend, newMessage);
+                sendingMessages.remove(messageToSend);
+                return;
+            }
 
-                MessageSending messageToSend = findMessageSendingMirror(message);
-                if (messageToSend != null) {
-                    messageListener.messageChanged(messageToSend, message);
-                    sendingMessages.remove(messageToSend);
-                } else {
-                    messageListener.messageAdded(sendingMessages.isEmpty()
-                            ? null
-                            : sendingMessages.get(0), message);
+            if (addToEnd) {
+                if (headMessage == null) {
+                    headMessage = newMessage;
                 }
+                currentMessages.add(newMessage);
+                onCurrentMessageAdded(null, newMessage);
+            } else {
+                if (previousChatLastMessageIndex > currentIndex) {
+                    previousChatLastMessageIndex++;
+                }
+
+                currentMessages.add(currentIndex, newMessage);
+                MessageImpl messageBefore = currentMessages.get(currentIndex + 1);
+                onCurrentMessageAdded(messageBefore, newMessage);
             }
         }
 
         @Nullable
-        private MessageSending findMessageSendingMirror(MessageImpl msg) {
-            for (MessageSending ms : sendingMessages) {
-                if (ms.getId().equals(msg.getId())) {
-                    return ms;
+        private MessageSending findMessageSendingMirror(MessageImpl newMessage) {
+            for (MessageSending sendingMessage : sendingMessages) {
+                if (sendingMessage.equals(newMessage)) {
+                    return sendingMessage;
                 }
             }
             return null;
         }
 
-        void onCurrentChatMessageChanged(int index,
-                                         @CurrentChat MessageImpl from,
-                                         @CurrentChat MessageImpl to) {
-            from.getSource().assertCurrentChat();
-            to.getSource().assertCurrentChat();
-            if (headMessage != null && headMessage.getSource().isHistory()
-                    || index >= currentChatMessages.indexOf(headMessage)) {
-                if (from == headMessage) {
-                    headMessage = to;
-                }
+        void onCurrentMessageAdded(MessageImpl before, MessageImpl msg) {
+            if (headMessage != null && msg.compareTo(headMessage) >= 0) {
+                messageListener.messageAdded(before, msg);
+            }
+        }
+
+        void onCurrentMessageChanged( MessageImpl from, MessageImpl to) {
+            if (headMessage != null && to.compareTo(headMessage) >= 0) {
                 messageListener.messageChanged(from, to);
             }
         }
 
-        void onCurrentChatMessageDeleted(int index, @CurrentChat MessageImpl msg) {
-            msg.getSource().assertCurrentChat();
-            if (headMessage != null) {
-                int headIndex = currentChatMessages.indexOf(headMessage);
-                if (headMessage.getSource().isHistory() || index > headIndex) {
-                    if (index + 1 == headIndex) {
-                        headMessage = currentChatMessages.size() < headIndex
-                                ? null
-                                : currentChatMessages.get(headIndex);
-                    } else {
-                        if (!currentChatMessages.isEmpty() && headIndex == -1) {
-                            int offset = (currentChatMessages.size() > index) ? 0 : 1;
-                            headMessage = currentChatMessages.get(index - offset);
-                        }
-                    }
-                }
+        void onCurrentMessageDeleted(int nextMessageIndex, MessageImpl msg) {
+            if (headMessage != null && msg.compareTo(headMessage) >= 0) {
                 messageListener.messageRemoved(msg);
-            }
-        }
 
-        void onHistoryChanged(@NonNull MessageImpl msg) {
-            msg.getSource().assertHistory();
-            if (headMessage != null) {
-                if (msg.getTimeMicros() >= headMessage.getTimeMicros()) {
-                    idToHistoryMessageMap.put(msg.getHistoryId().getDbId(), msg);
-                } else {
-                    setSecondaryHistoryToCurrentMessage(msg);
+                if (headMessage == msg) {
+                    headMessage = getNextCurrentMessage(nextMessageIndex);
                 }
             }
         }
 
-        void onHistoryDeleted(String id) {
-            MessageImpl msg = idToHistoryMessageMap.remove(id);
-            if (headMessage != null
-                    && headMessage.getSource().isHistory()
-                    && msg != null && msg.getTimeMicros() >= headMessage.getTimeMicros()) {
-                messageListener.messageRemoved(msg);
-            }
+        void onHistoryAdded(String beforeId, MessageImpl historyMessage) {
+            addNewOrMergeMessage(historyMessage);
         }
 
-        void onHistoryAdded(@Nullable HistoryId before, MessageImpl msg) {
-            msg.getSource().assertHistory();
-            if (headMessage != null && headMessage.getSource().isHistory()) {
-                if (before != null) {
-                    MessageImpl beforeMsg = idToHistoryMessageMap.get(before.getDbId());
-                    if (beforeMsg != null) {
-                        idToHistoryMessageMap.put(msg.getHistoryId().getDbId(), msg);
-                        messageListener.messageAdded(beforeMsg, msg);
+        void onHistoryChanged(@NonNull MessageImpl historyMessage) {
+//            TODO: Remove comments when task https://bugz.olegb.ru/browse/BUGZTASKS-6661 will be solved
+//            addNewOrMergeMessage(historyMessage);
+        }
+
+        void onHistoryDeleted(String serverSideId) {
+            for (int i = 0; i < currentMessages.size(); i++) {
+                MessageImpl chatMessage = currentMessages.get(i);
+                if (chatMessage.getServerSideId().equals(serverSideId)) {
+                    currentMessages.remove(i);
+                    onCurrentMessageDeleted(i, chatMessage);
+                    if (previousChatLastMessageIndex >= i) {
+                        // We deleted message from old chat, so decrement previousChatLastMessageIndex
+                        previousChatLastMessageIndex--;
                     }
-                } else {
-                    idToHistoryMessageMap.put(msg.getHistoryId().getDbId(), msg);
-                    messageListener.messageAdded(
-                            currentChatMessages.size() != 0
-                                    ? currentChatMessages.get(0)
-                                    : null,
-                            msg);
+                    break;
                 }
             }
         }
 
         void onHistoryEndOfBatch() {
-            if (!isFirstHistoryUpdateReceived) {
-                isFirstHistoryUpdateReceived = true;
+            if (!isFirstUpdateReceived) {
+                isFirstUpdateReceived = true;
                 if (cachedCallback != null) {
                     GetMessagesCallback callback = cachedCallback;
                     cachedCallback = null;
                     uncheckedGetNextMessages(cachedLimit, callback);
-                }
-            }
-        }
-
-        void setSecondaryHistoryToCurrentMessage(MessageImpl historyMessage) {
-            if (headMessage != null && historyMessage.getTimeMicros() < headMessage.getTimeMicros()) {
-                for (MessageImpl chatMessage : currentChatMessages) {
-                    if (historyMessage.getId().equals(chatMessage.getId())) {
-                        chatMessage.setSecondaryHistory(historyMessage);
-                        idToHistoryMessageMap.put(historyMessage.getHistoryId().getDbId(), historyMessage);
-                        break;
-                    }
                 }
             }
         }
@@ -979,49 +864,24 @@ public class MessageHolderImpl implements MessageHolder {
 
             @Override
             public void receive(@NonNull List<? extends Message> messages) {
+                if (isDestroyed) {
+                    return;
+                }
                 List<? extends Message> result;
                 if (!messages.isEmpty()) {
-                    if (!currentChatMessages.isEmpty()
-                            && messages.get(messages.size() - 1).getTime() >= currentChatMessages.get(0).getTime()) {
-                        // We received history that overlap current chat messages. Merging
-                        List<Message> filtered = new ArrayList<>(messages.size());
-                        MessageImpl first = (MessageImpl) messages.get(0);
-                        for (Message msg1 : messages) {
-                            MessageImpl msg = (MessageImpl) msg1;
-                            if (msg.getSource().isHistory()) {
-                                if (msg.getTime() >= currentChatMessages.get(0).getTime()
-                                        && msg.getTime() <= currentChatMessages.get(currentChatMessages.size() - 1).getTime()) {
-                                    for (MessageImpl cc : currentChatMessages) {
-                                        if (cc.getId().equals(msg.getId())) {
-                                            cc.setSecondaryHistory(msg);
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                            filtered.add(msg);
-                        }
-                        if (filtered.isEmpty()) {
-                            getMessages(first, limit, this);
-                            return;
-                        }
-                        result = filtered;
-                    } else {
-                        result = messages;
-                    }
+                    List<MessageImpl> filtered = new ArrayList<>(messages.size());
+                    merge((List<MessageImpl>) messages, filtered);
 
-                    for (Message msg : messages) {
-                        if (((MessageImpl) msg).getSource().isHistory()) {
-                            idToHistoryMessageMap.put(
-                                    ((MessageImpl) msg).getHistoryId().getDbId(),
-                                    (MessageImpl) msg
-                            );
+                    if (filtered.isEmpty()) {
+                        getMessages((MessageImpl) messages.get(0), limit, this);
+                        return;
+                    } else  {
+                        MessageImpl first = filtered.get(0);
+                        if (headMessage == null || first.compareTo(headMessage) < 0) {
+                            headMessage = first;
                         }
                     }
-                    MessageImpl first = (MessageImpl) result.get(0);
-                    if (headMessage == null || compare(first, headMessage) < 0) {
-                        headMessage = first;
-                    }
+                    result = filtered;
                 } else {
                     result = messages;
                     isAllMessageSourcesEnded = true;
@@ -1030,8 +890,34 @@ public class MessageHolderImpl implements MessageHolder {
                 wrapped.receive(unmodifiableList(result));
             }
 
-            private int compare(MessageImpl left, MessageImpl right) {
-                return InternalUtils.compare(left.getTimeMicros(), right.getTimeMicros());
+            private void merge(@NonNull List<MessageImpl> newMessages, @NonNull List<MessageImpl> filtered) {
+                int oldMessageIndex = 0;
+                for (MessageImpl newMessage : newMessages) {
+                    boolean merged = false;
+                    while (oldMessageIndex < currentMessages.size()) {
+                        MessageImpl currentMessage = currentMessages.get(oldMessageIndex);
+                        if (currentMessage.equals(newMessage)) {
+                            if (headMessage == null || newMessage.compareTo(headMessage) < 0) {
+                                filtered.add(newMessage);
+                            }
+                            merged = true;
+                            oldMessageIndex++;
+                            break;
+                        } else if (currentMessage.compareTo(newMessage) >= 0) {
+                            break;
+                        }
+                        oldMessageIndex++;
+                    }
+
+                    if (!merged) {
+                        currentMessages.add(oldMessageIndex, newMessage);
+                        filtered.add(newMessage);
+                        if (previousChatLastMessageIndex > oldMessageIndex) {
+                            previousChatLastMessageIndex++;
+                        }
+                        oldMessageIndex++;
+                    }
+                }
             }
         }
     }
