@@ -8,36 +8,42 @@ import com.webimapp.android.sdk.impl.items.ChatItem;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
+import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 
 @RunWith(MockitoJUnitRunner.class)
 public class MessageHolderImplTest {
 
     private Runnable EMPTY_RUNNABLE = () -> {};
 
-    private Set<String> EMPTY_SET = Collections.emptySet();
-
     private List<MessageImpl> EMPTY_MESSAGES_LIST = Collections.emptyList();
+
+    private <T> Set<T> setOf(T ... elements) {
+        return new HashSet<>(Arrays.asList(elements));
+    }
+
+    private <T> List<T> listOf(T ... elements) {
+        return Arrays.asList(elements);
+    }
 
     private ArgumentCaptor<List<Message>> newListCaptor() {
         return ArgumentCaptor.forClass(List.class);
@@ -59,8 +65,11 @@ public class MessageHolderImplTest {
             int before = -1;
             for (int i = history.size() - 1; i >= 0; i--) {
                 MessageImpl historyMessage = history.get(i);
-                if (historyMessage.getTimeMicros() <= beforeMessageTs) {
+                if (historyMessage.getTimeMicros() == beforeMessageTs) {
                     before = i;
+                    break;
+                } else if (historyMessage.getTimeMicros() < beforeMessageTs) {
+                    before = i + 1;
                     break;
                 }
             }
@@ -74,11 +83,9 @@ public class MessageHolderImplTest {
         }
     }
 
-    private int messageIndexCounter = 0;
-
-    private List<MessageImpl> generateHistory(int count) {
+    private List<MessageImpl> generateHistoryInRange(int first, int last) {
         List<MessageImpl> history = new ArrayList<>();
-        for (int i = messageIndexCounter; i < messageIndexCounter + count; i++) {
+        for (int i = first; i < last; i++) {
             history.add(new MessageImpl(
                 "",
                 StringId.forMessage(String.valueOf(i)),
@@ -103,13 +110,15 @@ public class MessageHolderImplTest {
                 null
             ));
         }
-        messageIndexCounter += count;
         return history;
     }
 
-    private List<MessageImpl> generateCurrentChat(int count) {
+    private List<MessageImpl> generateCurrentChatInRange(int first, int last, Set<Integer> expect) {
         List<MessageImpl> currentChatMessages = new ArrayList<>();
-        for (int i = messageIndexCounter; i < messageIndexCounter + count; i++) {
+        for (int i = first; i < last; i++) {
+            if (expect.contains(i)) {
+                continue;
+            }
             currentChatMessages.add(new MessageImpl(
                 "",
                 StringId.forMessage(String.valueOf(i)),
@@ -134,7 +143,6 @@ public class MessageHolderImplTest {
                 null
             ));
         }
-        messageIndexCounter += count;
         return currentChatMessages;
     }
 
@@ -144,7 +152,7 @@ public class MessageHolderImplTest {
             historyMessages.add(
                 new MessageImpl(
                     "",
-                    currentChatMessage.getId(),
+                    currentChatMessage.getClientSideId(),
                     currentChatMessage.getSessionId(),
                     currentChatMessage.getOperatorId(),
                     currentChatMessage.getAvatarUrlLastPart(),
@@ -170,19 +178,18 @@ public class MessageHolderImplTest {
         return historyMessages;
     }
 
-    private MessageImpl newCurrentChatMessage() {
-        int i = messageIndexCounter++;
+    private MessageImpl newCurrentChatMessage(int index) {
         return new MessageImpl(
             "",
-            StringId.forMessage(String.valueOf(i)),
+            StringId.forMessage(String.valueOf(index)),
             null,
             StringId.forOperator("op"),
             "",
             "",
             Message.Type.OPERATOR,
             "text",
-            i,
-            String.valueOf(i),
+            index,
+            String.valueOf(index),
             null,
             false,
             null,
@@ -200,7 +207,7 @@ public class MessageHolderImplTest {
     private MessageImpl newEditedCurrentChatMessage(MessageImpl original) {
         return new MessageImpl(
             "",
-            original.getId(),
+            original.getClientSideId(),
             original.getSessionId(),
             original.getOperatorId(),
             original.getAvatarUrlLastPart(),
@@ -208,7 +215,7 @@ public class MessageHolderImplTest {
             original.getType(),
             original.getText() + "1",
             original.getTimeMicros(),
-            original.getIdInCurrentChat(),
+            original.getServerSideId(),
             null,
             false,
             null,
@@ -226,7 +233,7 @@ public class MessageHolderImplTest {
     private MessageImpl newEditedHistoryMessage(MessageImpl original) {
         return new MessageImpl(
             "",
-            original.getId(),
+            original.getClientSideId(),
             original.getSessionId(),
             original.getOperatorId(),
             original.getAvatarUrlLastPart(),
@@ -234,7 +241,7 @@ public class MessageHolderImplTest {
             original.getType(),
             original.getText() + "1",
             original.getTimeMicros(),
-            original.getHistoryId().getDbId(),
+            original.getServerSideId(),
             null,
             true,
             null,
@@ -264,32 +271,22 @@ public class MessageHolderImplTest {
         return resultList;
     }
 
-    private void setCurrentChatMessages(MessageHolder messageHolder, List<MessageImpl> messages) {
+    private List<MessageImpl> cloneCurrentChatMessages(List<MessageImpl> currentMessages) {
         try {
-            Field currentChatMessages = messageHolder.getClass().getDeclaredField("currentChatMessages");
-            currentChatMessages.setAccessible(true);
-            currentChatMessages.set(messageHolder, messages);
-        } catch (Exception exception) {
-            exception.printStackTrace();
-        }
-    }
-
-    private List<MessageImpl> cloneCurrentChatMessages(List<MessageImpl> currentChatMessages) {
-        try {
-            Method method = currentChatMessages.getClass().getDeclaredMethod("clone");
+            Method method = currentMessages.getClass().getDeclaredMethod("clone");
             method.setAccessible(true);
-            return (List<MessageImpl>) method.invoke(currentChatMessages);
+            return (List<MessageImpl>) method.invoke(currentMessages);
         } catch (Exception exception) {
             exception.printStackTrace();
             return null;
         }
     }
 
-    private void resetLastChatMessageIndex(MessageHolder messageHolder) {
+    private void setPrivateField(String privateFieldName, Object privateFieldValue, Object object) {
         try {
-            Field lastChatMessageIndex = messageHolder.getClass().getDeclaredField("lastChatMessageIndex");
-            lastChatMessageIndex.setAccessible(true);
-            lastChatMessageIndex.set(messageHolder, 0);
+            Field field = object.getClass().getDeclaredField(privateFieldName);
+            field.setAccessible(true);
+            field.set(object, privateFieldValue);
         } catch (Exception exception) {
             exception.printStackTrace();
         }
@@ -297,16 +294,15 @@ public class MessageHolderImplTest {
 
     @Test
     public void internalHistoryGeneration() {
-        List<MessageImpl> history1 = generateHistory(10);
-        generateCurrentChat(10);
-        List<MessageImpl> history3 = generateHistory(10);
+        List<MessageImpl> history1 = generateHistoryInRange(0, 10);
+        List<MessageImpl> history3 = generateHistoryInRange(20, 30);
 
         List<String> ids1 = history1.stream()
-            .map(MessageImpl::getPrimaryId)
+            .map(MessageImpl::getServerSideId)
             .collect(Collectors.toList());
 
         List<String> ids3 = history3.stream()
-            .map(MessageImpl::getPrimaryId)
+            .map(MessageImpl::getServerSideId)
             .collect(Collectors.toList());
 
         assertEquals(ids1, IntStream.rangeClosed(0, 9).boxed().map(String::valueOf).collect(Collectors.toList()));
@@ -315,772 +311,231 @@ public class MessageHolderImplTest {
 
     @Test
     public void respondCurrentChatMessagesImmediately() {
-        List<MessageImpl> currentChatMessages = generateCurrentChat(10);
+        List<MessageImpl> currentChatMessages = generateCurrentChatInRange(0, 10, setOf());
         MessageHolder messageHolder = newMessageHolderWithHistory(new History(Collections.<MessageImpl>emptyList()));
-        setCurrentChatMessages(messageHolder, currentChatMessages);
         MessageListener messageListener = mock(MessageListener.class);
         MessageTracker messageTracker = messageHolder.newMessageTracker(messageListener);
         MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
+
+        messageHolder.onChatReceive(null, new ChatItem(), currentChatMessages);
 
         messageTracker.getNextMessages(10, callback);
         verify(callback, times(1)).receive(currentChatMessages);
     }
 
     @Test
-    public void awaitHistoryResponse() {
-        List<MessageImpl> history1 = generateHistory(10);
-        List<MessageImpl> history2 = generateHistory(10);
-        MessageHolder holder = newMessageHolderWithHistory(new History(concat(history1, history2)));
+    public void checkGetNextMessages() {
+        List<MessageImpl> currentChatMessages = generateCurrentChatInRange(0, 10, setOf());
+        List<MessageImpl> history = generateHistoryFromCurrentChat(currentChatMessages);
+        MessageHolder messageHolder = newMessageHolderWithHistory(new History(listOf()));
         MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
+        MessageTracker messageTracker = messageHolder.newMessageTracker(messageListener);
         MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
 
-        // request 10 messages
-        tracker.getNextMessages(10, callback);
-        // no callback called because history has not been received; callback was cached an will be called on history receive
-        verify(callback, never()).receive(EMPTY_MESSAGES_LIST);
+        // This method caches callback
+        messageTracker.getNextMessages(10, callback);
+        // Callback will be cached because local storage is empty.
+        verify(callback, never()).receive(Mockito.any());
+
         reset(callback);
-
-        // history receive
-        holder.receiveHistoryUpdate(history2, EMPTY_SET, EMPTY_RUNNABLE);
-        // called previously cached callback
-        verify(callback, times(1)).receive(history2);
-        reset(callback);
-
-        // request next 10 messages
-        tracker.getNextMessages(10, callback);
-        // received older history
-        verify(callback, times(1)).receive(history1);
-    }
-
-    @Test
-    public void awaitHistoryResponseWithCurrentChat() {
-        List<MessageImpl> history1 = generateHistory(10);
-        List<MessageImpl> history2 = generateHistory(10);
-        List<MessageImpl> currentChatMessages = generateCurrentChat(10);
-        MessageHolder holder = newMessageHolderWithHistory(new History(concat(history1, history2)));
-        setCurrentChatMessages(holder, currentChatMessages);
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-
-        tracker.getNextMessages(currentChatMessages.size(), callback);
+        messageHolder.onChatReceive(null, new ChatItem(), currentChatMessages);
+        // Our cached callback will be fired.
         verify(callback, times(1)).receive(currentChatMessages);
-        reset(callback);
 
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(0)).receive(EMPTY_MESSAGES_LIST);
-        reset(callback);
-
-        holder.receiveHistoryUpdate(history2, EMPTY_SET, EMPTY_RUNNABLE);
-        verify(callback, times(1)).receive(history2);
-        reset(callback);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(history1);
-        reset(callback);
+        reset(messageListener);
+        // History messages will be saved to local storage and merged with currentMessages
+        messageHolder.receiveHistoryUpdate(history, setOf(), EMPTY_RUNNABLE);
+        verify(messageListener, times(currentChatMessages.size())).messageChanged(Mockito.any(), Mockito.any());
     }
 
     @Test
-    public void insertHistoryBetweenOlderHistoryAndCurrentChat() {
-        List<MessageImpl> history1 = generateHistory(10);
-        List<MessageImpl> history2 = generateHistory(2);
-        List<MessageImpl> currentChatMessages = generateCurrentChat(10);
-        MessageHolder messageHolder = newMessageHolderWithHistory(new History(concat(history1, history2)));
+    public void checkGetAllMessages() {
+        List<MessageImpl> history1 = generateHistoryInRange(0, 10);
+        List<MessageImpl> currentChatMessages = generateCurrentChatInRange(10, 20, setOf());
+        List<MessageImpl> history2 = generateHistoryFromCurrentChat(currentChatMessages);
+        MessageHolder messageHolder = newMessageHolderWithHistory(new History(listOf()));
         MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = messageHolder.newMessageTracker(messageListener);
+        MessageTracker messageTracker = messageHolder.newMessageTracker(messageListener);
         MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ArgumentCaptor<List<Message>> listCaptor;
-        ArgumentCaptor<Message> messageCaptor;
 
-        messageHolder.receiveHistoryUpdate(history1, EMPTY_SET, EMPTY_RUNNABLE);
+        // This method doesn't cache callback, it just gets all messages from local storage
+        messageTracker.getAllMessages(callback);
+        // Local storage is empty, so we get empty list
+        verify(callback, times(1)).receive(EMPTY_MESSAGES_LIST);
+
+        messageHolder.onChatReceive(null, new ChatItem(), currentChatMessages);
+        // history1 will be saved to local storage
+        messageHolder.receiveHistoryUpdate(history1, setOf(), EMPTY_RUNNABLE);
+        // history2 will be saved to local storage
+        messageHolder.receiveHistoryUpdate(history2, setOf(), EMPTY_RUNNABLE);
+
+        reset(callback);
+        messageTracker.getAllMessages(callback);
+        verify(callback, times(1)).receive(concat(history1, history2));
+    }
+
+    @Test
+    public void checkGetLastMessages() {
+        List<MessageImpl> currentChatMessages = generateCurrentChatInRange(0, 10, setOf());
+        List<MessageImpl> history1 = generateHistoryFromCurrentChat(currentChatMessages);
+        List<MessageImpl> history2 = generateHistoryInRange(10, 20);
+        MessageHolder messageHolder = newMessageHolderWithHistory(new History(Collections.<MessageImpl>emptyList()));
+        MessageListener messageListener = mock(MessageListener.class);
+        MessageTracker messageTracker = messageHolder.newMessageTracker(messageListener);
+        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
+
+        // This method caches callback
+        messageTracker.getLastMessages(10, callback);
+        // Local storage is empty, so callback is cached and will not be fired
+        verify(callback, never()).receive(Mockito.any());
+
+        reset(callback);
+        messageHolder.onChatReceive(null, new ChatItem(), currentChatMessages);
+        // Fire cached callback
+        verify(callback, times(1)).receive(currentChatMessages);
+
+        reset(messageListener);
+        messageHolder.receiveHistoryUpdate(history1, setOf(), EMPTY_RUNNABLE);
+        verify(messageListener, times(history1.size())).messageChanged(Mockito.any(), Mockito.any());
+        messageHolder.receiveHistoryUpdate(history2, setOf(), EMPTY_RUNNABLE);
+        verify(messageListener, times(history2.size())).messageAdded(Mockito.isNull(), Mockito.any());
+
+        reset(callback);
+        messageTracker.getLastMessages(10, callback);
+        verify(callback, times(1)).receive(history2);
+    }
+
+    @Test
+    public void checkResetTo() {
+        List<MessageImpl> currentChatMessages = generateCurrentChatInRange(0, 10, setOf());
+        MessageHolder messageHolder = newMessageHolderWithHistory(new History(Collections.<MessageImpl>emptyList()));
+        MessageListener messageListener = mock(MessageListener.class);
+        MessageTracker messageTracker = messageHolder.newMessageTracker(messageListener);
+        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
+
         messageHolder.onChatReceive(null, new ChatItem(), currentChatMessages);
 
-        listCaptor = newListCaptor();
-        tracker.getNextMessages(10, callback);
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(2)).receive(listCaptor.capture());
-        assertEquals(currentChatMessages, listCaptor.getAllValues().get(0));
-        assertEquals(history1, listCaptor.getAllValues().get(1));
-        reset(callback);
+        messageTracker.getNextMessages(8, callback);
+        verify(callback, times(1)).receive(currentChatMessages.subList(2, 10));
 
-        messageCaptor = newMessageCaptor();
-        // Receive history between current chat and older received history
-        messageHolder.receiveHistoryUpdate(history2, EMPTY_SET, EMPTY_RUNNABLE);
-        verify(messageListener, times(2)).messageAdded(messageCaptor.capture(), messageCaptor.capture());
-        // first history message inserted before first current chat message
-        assertEquals(currentChatMessages.get(0), messageCaptor.getAllValues().get(0));
-        assertEquals(history2.get(0), messageCaptor.getAllValues().get(1));
-        // second history message inserted before first current chat message
-        assertEquals(currentChatMessages.get(0), messageCaptor.getAllValues().get(2));
-        assertEquals(history2.get(1), messageCaptor.getAllValues().get(3));
-    }
-
-    @Test
-    public void receiveHistoryFirstPartOfCurrentChatWhenCurrentChatAndPreviousHistoryAreTracked() {
-        List<MessageImpl> history1 = generateHistory(10);
-        List<MessageImpl> currentChatMessages = generateCurrentChat(10);
-        List<MessageImpl> history2 = generateHistoryFromCurrentChat(currentChatMessages);
-        MessageHolder holder = newMessageHolderWithHistory(new History(concat(history1, history2)));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-
-        holder.receiveHistoryUpdate(history1, EMPTY_SET, EMPTY_RUNNABLE);
-        holder.onChatReceive(null, new ChatItem(), currentChatMessages);
-
-        // Request current chat
-        tracker.getNextMessages(10, callback);
-        // Received current chat
-        verify(callback, times(1)).receive(currentChatMessages);
-        reset(callback);
-
-        // Request history part
-        tracker.getNextMessages(5, callback);
-        // Received history part
-        verify(callback, times(1)).receive(history1.subList(5, 10));
-        reset(callback);
-
-        // Receive history part of current chat
-        holder.receiveHistoryUpdate(history2.subList(0, 6), EMPTY_SET, EMPTY_RUNNABLE);
-        // No any callbacks called
-        verify(callback, never()).receive(EMPTY_MESSAGES_LIST);
-        verifyNoInteractions(messageListener);
-        reset(callback);
-
-        // Request remaining history
-        tracker.getNextMessages(5, callback);
-        verify(callback, times(1)).receive(history1.subList(0, 5));
-    }
-
-    @Test
-    public void receiveFullHistoryOfCurrentChatWhenOnlyCurrentChatIsTracked() {
-        List<MessageImpl> history1 = generateHistory(10);
-        List<MessageImpl> currentChat = generateCurrentChat(10);
-        List<MessageImpl> history2 = generateHistoryFromCurrentChat(currentChat);
-        MessageHolder holder = newMessageHolderWithHistory(new History(concat(history1, history2)));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ArgumentCaptor<List<Message>> listCaptor;
-
-        holder.receiveHistoryUpdate(history1, EMPTY_SET, EMPTY_RUNNABLE);
-        holder.onChatReceive(null, new ChatItem(), currentChat);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(currentChat);
-        reset(callback);
-
-        holder.receiveHistoryUpdate(history2.subList(0, 6), EMPTY_SET, EMPTY_RUNNABLE);
-        verify(callback, never()).receive(EMPTY_MESSAGES_LIST);
-        verifyNoInteractions(messageListener);
-        reset(callback);
-
-        listCaptor = newListCaptor();
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(listCaptor.capture());
-        List<Message> callbackResult = listCaptor.getValue();
-        assertFalse(callbackResult.isEmpty());
-        assertEquals(history1.subList(10 - callbackResult.size(), 10), callbackResult);
-    }
-
-    @Test
-    public void receivedHistoryPartOfCurrentChatWhenOnlyCurrentChatIsTracked() {
-        List<MessageImpl> history1 = generateHistory(10);
-        List<MessageImpl> currentChat = generateCurrentChat(10);
-        List<MessageImpl> history2 = generateHistoryFromCurrentChat(currentChat);
-        MessageHolder holder = newMessageHolderWithHistory(new History(concat(history1, history2)));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ChatItem firstChat = new ChatItem();
-        ArgumentCaptor<List<Message>> listCaptor;
-
-        holder.receiveHistoryUpdate(history1, EMPTY_SET, EMPTY_RUNNABLE);
-        holder.onChatReceive(null, firstChat, currentChat.subList(0, 6));
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(currentChat.subList(0, 6));
-        reset(callback);
-
-        holder.receiveHistoryUpdate(history2.subList(0, 6), EMPTY_SET, EMPTY_RUNNABLE);
-        verifyNoInteractions(messageListener);
-
-        holder.receiveHistoryUpdate(history2.subList(6, 7), EMPTY_SET, EMPTY_RUNNABLE);
-        verifyNoInteractions(messageListener);
-
-        holder.receiveNewMessage(currentChat.get(6));
-        verify(messageListener, times(1)).messageAdded(null, currentChat.get(6));
+        // Now headMessage is set to message with index '2', let's reset currentMessages to '5'
         reset(messageListener);
+        messageTracker.resetTo(currentChatMessages.get(5));
+        // Check that for messages from '2' to '5' was called callback
+        verify(messageListener, times(3)).messageRemoved(any());
 
-        listCaptor = newListCaptor();
-        tracker.getNextMessages(5, callback);
-        verify(callback, times(1)).receive(listCaptor.capture());
-        List<Message> result = listCaptor.getValue();
-        assertFalse(result.isEmpty());
-        assertEquals(history1.subList(10 - result.size(), 10), result);
-    }
-
-    @Test
-    public void receiveCurrentChatWhenItsFullPartOfHistoryIsTracked() {
-        List<MessageImpl> history1 = generateHistory(10);
-        List<MessageImpl> currentChat = generateCurrentChat(10);
-        List<MessageImpl> history2 = generateHistoryFromCurrentChat(currentChat);
-        MessageHolder holder = newMessageHolderWithHistory(new History(concat(history1, history2)));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ChatItem firstChat = new ChatItem();
-
-        holder.receiveHistoryUpdate(concat(history1, history2), EMPTY_SET, EMPTY_RUNNABLE);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(history2);
         reset(callback);
+        messageTracker.getNextMessages(5, callback);
+        // Here we must get messages from '5' to '0'
+        verify(callback, times(1)).receive(currentChatMessages.subList(0, 5));
 
-        holder.onChatReceive(null, firstChat, currentChat.subList(0, 9));
-        verifyNoInteractions(messageListener);
-
-        holder.receiveNewMessage(currentChat.get(9));
-        verifyNoInteractions(messageListener);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(history1);
-    }
-
-    @Test
-    public void sequentiallyReceiveLocalHistoryToRemoteHistoryToCurrentChat() {
-        List<MessageImpl> history1 = generateHistory(10);
-        List<MessageImpl> currentChat = generateCurrentChat(10);
-        List<MessageImpl> history2 = generateHistoryFromCurrentChat(currentChat);
-        MessageHolder holder = newMessageHolderWithHistory(
-            new History(concat(history1, history2)),
-            concat(history1, history2.subList(0, 8))
-        );
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ChatItem firstChat = new ChatItem();
-
-        tracker.getNextMessages(8, callback);
-        verify(callback, times(1)).receive(history2.subList(0, 8));
-        reset(callback);
-
-        holder.receiveHistoryUpdate(concat(history1, history2.subList(8, 9)), EMPTY_SET, EMPTY_RUNNABLE);
-        verify(messageListener, times(1))
-            .messageAdded(null, history2.get(8));
         reset(messageListener);
+        messageTracker.resetTo(currentChatMessages.get(9));
+        verify(messageListener, times(9)).messageRemoved(any());
 
-        holder.onChatReceive(null, firstChat, currentChat);
-        verify(messageListener, times(1))
-            .messageAdded(null, currentChat.get(9));
+        reset(callback);
+        messageTracker.getNextMessages(10, callback);
+        verify(callback, times(1)).receive(currentChatMessages.subList(0, 9));
+    }
+
+    @Test
+    public void checkMergeCurrentChatWith() {
+        List<MessageImpl> currentChatMessages1 = generateCurrentChatInRange(0,10, setOf(1, 2, 3));
+        List<MessageImpl> currentChatMessages2 = generateCurrentChatInRange(0,10, setOf(8));
+        MessageHolder messageHolder = newMessageHolderWithHistory(new History(Collections.<MessageImpl>emptyList()));
+        MessageListener messageListener = mock(MessageListener.class);
+        MessageTracker messageTracker = messageHolder.newMessageTracker(messageListener);
+        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
+
+        ChatItem chatItem = new ChatItem();
+        // set these private fields, to make chat items equals
+        setPrivateField("id", "1id", chatItem);
+        setPrivateField("clientSideId", "1csi", chatItem);
+
+        assertEquals(chatItem, chatItem);
+
+        messageHolder.onChatReceive(null, chatItem, currentChatMessages1);
+        messageTracker.getNextMessages(10, callback);
+        verify(callback, times(1)).receive(currentChatMessages1);
+
         reset(messageListener);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(history1);
+        // Here we merging to equals chats with different currentChatMessages
+        messageHolder.onChatReceive(chatItem, chatItem, currentChatMessages2);
+        verify(messageListener, times(3)).messageAdded(any(), any()); // for 1, 2, 3
+        verify(messageListener, times(1)).messageRemoved(any()); // for 8
     }
 
     @Test
-    public void receiveCurrentChatWhenItsLastPartOfHistoryIsTracked() {
-        List<MessageImpl> currentChatMessages = generateCurrentChat(10);
-        List<MessageImpl> history2 = generateHistoryFromCurrentChat(currentChatMessages);
-        MessageHolder holder = newMessageHolderWithHistory(new History(history2));
+    public void checkMessageAddedCallbackFired() {
+        List<MessageImpl> currentMessages = generateCurrentChatInRange(0,10, setOf(8));
+        MessageHolder messageHolder = newMessageHolderWithHistory(new History(Collections.<MessageImpl>emptyList()));
         MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
+        MessageTracker messageTracker = messageHolder.newMessageTracker(messageListener);
         MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ChatItem firstChat = new ChatItem();
 
-        holder.receiveHistoryUpdate(history2.subList(1, 10), EMPTY_SET, EMPTY_RUNNABLE);
+        messageHolder.onChatReceive(null, new ChatItem(), currentMessages);
+        // call this method to set MessageTracker#headMessage pointer first current message '0',
+        // because MessageListener fires callbacks only if message placed before headMessage
+        messageTracker.getNextMessages(10, callback);
 
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(history2.subList(1, 10));
-        reset(callback);
-
-        holder.onChatReceive(null, firstChat, currentChatMessages);
-        verifyNoInteractions(messageListener);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(currentChatMessages.subList(0, 1));
-    }
-
-    @Test
-    public void receivingChatWithExistingMessagesAndChatMerge() {
-        MessageHolder holder = newMessageHolderWithHistory(new History(EMPTY_MESSAGES_LIST));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ArgumentCaptor<Message> messageCaptor;
-
-        ChatItem firstChat = new ChatItem();
-        List<MessageImpl> messages = generateCurrentChat(10);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, never()).receive(EMPTY_MESSAGES_LIST);
-
-        holder.onChatReceive(null, firstChat, messages.subList(0, 2));
-        verify(callback, times(1)).receive(messages.subList(0, 2));
-        verify(messageListener, never()).messageAdded(any(Message.class), any(Message.class));
-        reset(callback);
-
-        holder.onChatReceive(firstChat, firstChat, messages.subList(0, 2));
-        verifyNoInteractions(messageListener);
-
-        messageCaptor = newMessageCaptor();
-        holder.onChatReceive(firstChat, firstChat, messages.subList(0, 4));
-        verify(messageListener, times(2)).messageAdded(messageCaptor.capture(), messageCaptor.capture());
-        verify(messageListener, never()).messageRemoved(any(Message.class));
-        verify(messageListener, never()).messageChanged(any(Message.class), any(Message.class));
-        assertNull(messageCaptor.getAllValues().get(0));
-        assertEquals(messages.get(2), messageCaptor.getAllValues().get(1));
-        assertNull(messageCaptor.getAllValues().get(2));
-        assertEquals(messages.get(3), messageCaptor.getAllValues().get(3));
         reset(messageListener);
+        MessageImpl newMessage = newCurrentChatMessage(11);
+        // Add message to end of chat
+        messageHolder.onMessageAdded(newMessage);
+        // Verify message added to end
+        verify(messageListener, times(1)).messageAdded(null, newMessage);
 
-        holder.onChatReceive(firstChat, firstChat, concat(messages.subList(0, 2), messages.subList(3, 5)));
-        verify(messageListener, times(1)).messageRemoved(messages.get(2));
-        verify(messageListener, times(1)).messageAdded(null, messages.get(4));
-    }
-
-    @Test
-    public void replacingCurrentChat() {
-        MessageHolder holder = newMessageHolderWithHistory(new History(EMPTY_MESSAGES_LIST));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ArgumentCaptor<Message> messageCaptor;
-
-        ChatItem firstChat = new ChatItem("1");
-        ChatItem secondChat = new ChatItem("2");
-        List<MessageImpl> messages = generateCurrentChat(10);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, never()).receive(EMPTY_MESSAGES_LIST);
-
-        holder.onChatReceive(null, firstChat, messages.subList(0, 2));
-        verify(callback, times(1)).receive(messages.subList(0, 2));
-        verifyNoInteractions(messageListener);
-        reset(callback);
-
-        messageCaptor = newMessageCaptor();
-        holder.onChatReceive(firstChat, secondChat, messages.subList(2, 5));
-        verify(messageListener, times(3)).messageAdded(messageCaptor.capture(), messageCaptor.capture());
-        assertNull(messageCaptor.getAllValues().get(0));
-        assertEquals(messages.get(2), messageCaptor.getAllValues().get(1));
-        assertNull(messageCaptor.getAllValues().get(2));
-        assertEquals(messages.get(3), messageCaptor.getAllValues().get(3));
-        assertNull(messageCaptor.getAllValues().get(4));
-        assertEquals(messages.get(4), messageCaptor.getAllValues().get(5));
         reset(messageListener);
+        newMessage = newCurrentChatMessage(8); // message with id '8'
+        // Add message to middle of chat
+        messageHolder.onMessageAdded(newMessage);
+        // Verify message added to middle
+        MessageImpl beforeMessage = currentMessages.get(8); // message with id '9'
+        verify(messageListener, times(1)).messageAdded(beforeMessage, newMessage);
+    }
 
-        tracker.resetTo(messages.get(4));
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(messages.subList(0, 4));
-        reset(callback);
 
-        holder.onChatReceive(secondChat, secondChat, concat(messages.subList(2, 3), messages.subList(4, 6)));
-        verify(messageListener, times(1)).messageRemoved(messages.get(3));
-        verify(messageListener, times(1)).messageAdded(null, messages.get(5));
+    @Test
+    public void checkMessageChangedCallbackFired() {
+        List<MessageImpl> currentMessages = generateCurrentChatInRange(0,10, setOf());
+        MessageHolder messageHolder = newMessageHolderWithHistory(new History(Collections.<MessageImpl>emptyList()));
+        MessageListener messageListener = mock(MessageListener.class);
+        MessageTracker messageTracker = messageHolder.newMessageTracker(messageListener);
+        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
+
+        messageHolder.onChatReceive(null, new ChatItem(), currentMessages);
+        // call this method to set MessageTracker#headMessage pointer first current message '0',
+        // because MessageListener fires callbacks only if message placed before headMessage
+        messageTracker.getNextMessages(10, callback);
+
         reset(messageListener);
-
-        tracker.resetTo(messages.get(5));
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(concat(messages.subList(0, 3), messages.subList(4, 5)));
+        MessageImpl originalMessage = currentMessages.get(8);
+        MessageImpl editedMessage = newEditedHistoryMessage(originalMessage);
+        // Add message to middle of chat
+        messageHolder.onMessageChanged(editedMessage);
+        // Verify message was changed at position '8'
+        verify(messageListener, times(1)).messageChanged(originalMessage, editedMessage);
     }
 
+
     @Test
-    public void replacingCurrentChatWhenPreviousChatHistoryAlreadyReceived() {
-        List<MessageImpl> history1 = generateHistory(10);
-        List<MessageImpl> messages = generateCurrentChat(10);
-        List<MessageImpl> history2 = generateHistoryFromCurrentChat(messages);
-        MessageHolder holder = newMessageHolderWithHistory(new History(concat(history1, history2)));
+    public void checkMessageDeletedCallbackFired() {
+        List<MessageImpl> currentMessages = generateCurrentChatInRange(0,10, setOf());
+        MessageHolder messageHolder = newMessageHolderWithHistory(new History(Collections.<MessageImpl>emptyList()));
         MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
+        MessageTracker messageTracker = messageHolder.newMessageTracker(messageListener);
         MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ArgumentCaptor<Message> messageCaptor;
 
-        ChatItem firstChat = new ChatItem("1");
-        ChatItem secondChat = new ChatItem("2");
+        messageHolder.onChatReceive(null, new ChatItem(), currentMessages);
+        // call this method to set MessageTracker#headMessage pointer first current message '0',
+        // because MessageListener fires callbacks only if message placed before headMessage
+        messageTracker.getNextMessages(10, callback);
 
-        tracker.getNextMessages(10, callback);
-        verify(callback, never()).receive(EMPTY_MESSAGES_LIST);
-        reset(callback);
-
-        holder.onChatReceive(null, firstChat, messages.subList(0, 2));
-        verify(callback, times(1)).receive(messages.subList(0, 2));
-        verifyNoInteractions(messageListener);
-        reset(callback);
-
-        holder.receiveHistoryUpdate(history2.subList(0, 2), EMPTY_SET, EMPTY_RUNNABLE);
-        verifyNoInteractions(messageListener);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(history1);
-        reset(callback);
-
-        messageCaptor = newMessageCaptor();
-        holder.onChatReceive(firstChat, secondChat, messages.subList(2, 5));
-        verify(messageListener, times(3)).messageAdded(messageCaptor.capture(), messageCaptor.capture());
-        assertNull(messageCaptor.getAllValues().get(0));
-        assertEquals(messages.get(2), messageCaptor.getAllValues().get(1));
-        assertNull(messageCaptor.getAllValues().get(2));
-        assertEquals(messages.get(3), messageCaptor.getAllValues().get(3));
-        assertNull(messageCaptor.getAllValues().get(4));
-        assertEquals(messages.get(4), messageCaptor.getAllValues().get(5));
-        resetLastChatMessageIndex(holder);
         reset(messageListener);
-
-        tracker.resetTo(messages.get(4));
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(messages.subList(2, 4));
-        reset(callback);
-
-        tracker.getNextMessages(20, callback);
-        verify(callback, times(1)).receive(concat(history1, history2.subList(0, 2)));
-        reset(callback);
-
-        holder.onChatReceive(secondChat, secondChat, concat(messages.subList(2, 3), messages.subList(4, 6)));
-        verify(messageListener, times(1)).messageRemoved(messages.get(3));
-        reset(messageListener);
-
-        tracker.resetTo(messages.get(5));
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(concat(messages.subList(2, 3), messages.subList(4, 5)));
-    }
-
-    @Test
-    public void replacingCurrentChatWhenPreviousChatHistoryStoredLocally() {
-        List<MessageImpl> history1 = generateHistory(10);
-        List<MessageImpl> messages = generateCurrentChat(10);
-        List<MessageImpl> history2 = generateHistoryFromCurrentChat(messages);
-        MessageHolder holder = newMessageHolderWithHistory(new History(concat(history1, history2)), history2.subList(0, 2));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ArgumentCaptor<Message> messageCaptor;
-
-        ChatItem firstChat = new ChatItem("1");
-        ChatItem secondChat = new ChatItem("2");
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(history2.subList(0, 2));
-        reset(callback);
-
-        holder.onChatReceive(null, firstChat, messages.subList(0, 2));
-        verify(callback, never()).receive(EMPTY_MESSAGES_LIST);
-        verifyNoInteractions(messageListener);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(history1);
-        reset(callback);
-
-        messageCaptor = newMessageCaptor();
-        holder.onChatReceive(firstChat, secondChat, messages.subList(2, 5));
-        verify(messageListener, times(3)).messageAdded(messageCaptor.capture(), messageCaptor.capture());
-        assertNull(messageCaptor.getAllValues().get(0));
-        assertEquals(messages.get(2), messageCaptor.getAllValues().get(1));
-        assertNull(messageCaptor.getAllValues().get(2));
-        assertEquals(messages.get(3), messageCaptor.getAllValues().get(3));
-        assertNull(messageCaptor.getAllValues().get(4));
-        assertEquals(messages.get(4), messageCaptor.getAllValues().get(5));
-        resetLastChatMessageIndex(holder);
-        reset(messageListener);
-
-        tracker.resetTo(messages.get(4));
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(messages.subList(2, 4));
-        reset(callback);
-
-        tracker.getNextMessages(20, callback);
-        verify(callback, times(1)).receive(concat(history1, history2.subList(0, 2)));
-        reset(callback);
-
-        holder.onChatReceive(secondChat, secondChat, concat(messages.subList(2, 3), messages.subList(4, 6)));
-        verify(messageListener, times(1)).messageRemoved(messages.get(3));
-        verify(messageListener, times(1)).messageAdded(null, messages.get(5));
-        reset(messageListener);
-
-        tracker.resetTo(messages.get(5));
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(concat(messages.subList(2, 3), messages.subList(4, 5)));
-    }
-
-    @Test
-    public void mixedCurrentChatAndHistoryReset() {
-        List<MessageImpl> history1 = generateHistory(10);
-        List<MessageImpl> history2 = generateHistory(10);
-        List<MessageImpl> currentChat = generateCurrentChat(10);
-        MessageImpl nextCurrentChatMessage = newCurrentChatMessage();
-        MessageHolder holder = newMessageHolderWithHistory(new History(concat(history1, history2)));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ChatItem firstChat = new ChatItem();
-        ArgumentCaptor<List<Message>> listCaptor;
-
-        holder.receiveHistoryUpdate(history2, EMPTY_SET, EMPTY_RUNNABLE);
-        holder.onChatReceive(null, firstChat, currentChat);
-
-        listCaptor = newListCaptor();
-        tracker.getNextMessages(10, callback);
-        tracker.getNextMessages(10, callback);
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(3)).receive(listCaptor.capture());
-        assertEquals(currentChat, listCaptor.getAllValues().get(0));
-        assertEquals(history2, listCaptor.getAllValues().get(1));
-        assertEquals(history1, listCaptor.getAllValues().get(2));
-        reset(callback);
-
-        holder.receiveNewMessage(nextCurrentChatMessage);
-        verify(messageListener, times(1)).messageAdded(null, nextCurrentChatMessage);
-        reset(messageListener);
-
-        listCaptor = newListCaptor();
-        tracker.resetTo(nextCurrentChatMessage);
-        tracker.getNextMessages(10, callback);
-        tracker.getNextMessages(10, callback);
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(3)).receive(listCaptor.capture());
-        assertEquals(currentChat, listCaptor.getAllValues().get(0));
-        assertEquals(history2, listCaptor.getAllValues().get(1));
-        assertEquals(history1, listCaptor.getAllValues().get(2));
-        reset(callback);
-    }
-
-    @Test
-    public void emptyHistoryAndReceiveNewMessage() {
-        MessageImpl nextCurrentChatMessage = newCurrentChatMessage();
-        MessageHolder holder = newMessageHolderWithHistory(new History(EMPTY_MESSAGES_LIST));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-
-        holder.receiveHistoryUpdate(EMPTY_MESSAGES_LIST, EMPTY_SET, EMPTY_RUNNABLE);
-        holder.setReachedEndOfRemoteHistory(true);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(EMPTY_MESSAGES_LIST);
-        reset(callback);
-
-        holder.receiveNewMessage(nextCurrentChatMessage);
-        verify(messageListener, times(1)).messageAdded(null, nextCurrentChatMessage);
-    }
-
-    @Test
-    public void currentChatMessageEdition() {
-        MessageImpl nextCurrentChatMessage = newCurrentChatMessage();
-        MessageImpl editedCurrentChatMessage = newEditedCurrentChatMessage(nextCurrentChatMessage);
-        MessageHolder holder = newMessageHolderWithHistory(new History(EMPTY_MESSAGES_LIST));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-
-        holder.receiveHistoryUpdate(EMPTY_MESSAGES_LIST, EMPTY_SET, EMPTY_RUNNABLE);
-        holder.setReachedEndOfRemoteHistory(true);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(EMPTY_MESSAGES_LIST);
-        reset(callback);
-
-        holder.receiveNewMessage(nextCurrentChatMessage);
-        verify(messageListener, times(1)).messageAdded(null, nextCurrentChatMessage);
-        reset(messageListener);
-
-        holder.onMessageChanged(editedCurrentChatMessage);
-        verify(messageListener, times(1)).messageChanged(nextCurrentChatMessage, editedCurrentChatMessage);
-    }
-
-    @Test
-    public void currentChatEditionAndReceivedWithFullUpdate() {
-        List<MessageImpl> currentChat = generateCurrentChat(10);
-        List<MessageImpl> editedCurrentChatMessage = cloneCurrentChatMessages(currentChat);
-        currentChat.set(9, newEditedCurrentChatMessage(currentChat.get(9)));
-        MessageHolder holder = newMessageHolderWithHistory(new History(EMPTY_MESSAGES_LIST));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ChatItem chat = new ChatItem();
-
-        holder.receiveHistoryUpdate(EMPTY_MESSAGES_LIST, EMPTY_SET, EMPTY_RUNNABLE);
-        holder.setReachedEndOfRemoteHistory(true);
-        holder.onChatReceive(null, chat, currentChat);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(currentChat);
-        reset(callback);
-
-        holder.onChatReceive(chat, chat, editedCurrentChatMessage);
-        verify(messageListener, times(1))
-            .messageChanged(currentChat.get(9), editedCurrentChatMessage.get(9));
-    }
-
-    @Test
-    public void replaceHistoryMessageWithEditedCurrentChat() {
-        List<MessageImpl> currentChatMessages = generateCurrentChat(10);
-        List<MessageImpl> history = generateHistoryFromCurrentChat(currentChatMessages);
-        currentChatMessages.set(9, newEditedCurrentChatMessage(currentChatMessages.get(9)));
-        MessageHolder holder = newMessageHolderWithHistory(new History(history));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-
-        holder.receiveHistoryUpdate(history, EMPTY_SET, EMPTY_RUNNABLE);
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(history);
-        reset(callback);
-
-        holder.onChatReceive(null, new ChatItem(), currentChatMessages);
-        verify(messageListener, times(1))
-            .messageChanged(history.get(9), currentChatMessages.get(9));
-    }
-
-    @Test
-    public void replaceCurrentChatMessageWithEditedHistory() {
-        // (chat closed -> history received -> history edited)
-        List<MessageImpl> currentChat = generateCurrentChat(10);
-        List<MessageImpl> history = generateHistoryFromCurrentChat(currentChat);
-        MessageImpl editedHistory = newEditedHistoryMessage(history.get(9));
-        MessageHolder holder = newMessageHolderWithHistory(new History(history));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ChatItem chat = new ChatItem();
-
-        holder.onChatReceive(null, chat, currentChat);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(currentChat);
-        reset(callback);
-
-        holder.onChatReceive(chat, null, EMPTY_MESSAGES_LIST);
-        holder.receiveHistoryUpdate(history, EMPTY_SET, EMPTY_RUNNABLE);
-        verify(messageListener, times(20)).messageChanged(any(Message.class), any(Message.class));
-        reset(messageListener);
-
-        holder.receiveHistoryUpdate(Collections.singletonList(editedHistory), EMPTY_SET, EMPTY_RUNNABLE);
-        verify(messageListener, never()).messageChanged(any(Message.class), any(Message.class));
-    }
-
-    @Test
-    public void replaceCurrentChatMessageWithEditedHistory2() {
-        // (chat still open -> history received -> history edited)
-        List<MessageImpl> currentChat = generateCurrentChat(10);
-        List<MessageImpl> history = generateHistoryFromCurrentChat(currentChat);
-        MessageImpl editedHistory = newEditedHistoryMessage(history.get(9));
-        MessageHolder holder = newMessageHolderWithHistory(new History(history));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ChatItem chat = new ChatItem();
-
-        holder.onChatReceive(null, chat, currentChat);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(currentChat);
-        reset(callback);
-
-        holder.receiveHistoryUpdate(history, EMPTY_SET, EMPTY_RUNNABLE);
-        verifyNoInteractions(messageListener);
-
-        holder.receiveHistoryUpdate(Collections.singletonList(editedHistory), EMPTY_SET, EMPTY_RUNNABLE);
-        verifyNoInteractions(messageListener);
-    }
-
-    @Test
-    public void replaceCurrentChatMessageWithEditedHistory3() {
-        // (chat closed -> received edited history)
-        List<MessageImpl> currentChat = generateCurrentChat(10);
-        List<MessageImpl> history = generateHistoryFromCurrentChat(currentChat);
-        history.set(9, newEditedHistoryMessage(history.get(9)));
-        MessageHolder holder = newMessageHolderWithHistory(new History(history));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ChatItem chat = new ChatItem();
-
-        holder.onChatReceive(null, chat, currentChat);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(currentChat);
-        reset(callback);
-
-        holder.onChatReceive(chat, null, EMPTY_MESSAGES_LIST);
-        verify(messageListener, times(10)).messageChanged(any(Message.class), any(Message.class));
-        reset(messageListener);
-
-        holder.receiveHistoryUpdate(history, EMPTY_SET, EMPTY_RUNNABLE);
-        verify(messageListener, times(1)).messageChanged(history.get(9), history.get(9));
-    }
-
-    @Test
-    public void replaceCurrentChatMessageWithEditedHistory4() {
-        // (received edited history -> chat closed)
-        List<MessageImpl> currentChat = generateCurrentChat(10);
-        List<MessageImpl> history = generateHistoryFromCurrentChat(currentChat);
-        history.set(9, newEditedHistoryMessage(history.get(9)));
-        MessageHolder holder = newMessageHolderWithHistory(new History(history));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ChatItem chat = new ChatItem();
-
-        holder.onChatReceive(null, chat, currentChat);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(currentChat);
-        reset(callback);
-
-        holder.receiveHistoryUpdate(history, EMPTY_SET, EMPTY_RUNNABLE);
-        verifyNoInteractions(messageListener);
-
-        holder.onChatReceive(chat, null, EMPTY_MESSAGES_LIST);
-        verify(messageListener, times(1)).messageChanged(currentChat.get(9), history.get(9));
-    }
-
-    @Test
-    public void receivingCurrentChatMessageWhenItIsHeldAsLocalHistory() {
-        List<MessageImpl> currentChat = generateCurrentChat(10);
-        List<MessageImpl> history = generateHistoryFromCurrentChat(currentChat);
-        MessageHolder holder = newMessageHolderWithHistory(new History(history));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-
-        holder.receiveHistoryUpdate(history, EMPTY_SET, EMPTY_RUNNABLE);
-
-        tracker.getNextMessages(10, callback);
-        verify(callback, times(1)).receive(history);
-        reset(callback);
-
-        holder.receiveNewMessage(currentChat.get(9));
-        holder.receiveNewMessage(currentChat.get(9));
-        verify(callback, never()).receive(EMPTY_MESSAGES_LIST);
-    }
-
-    @Test
-    public void replacingHistoryWithCurrentChatDeletingMessagesReplacingItWithChatAgain() {
-        List<MessageImpl> currentChat = generateCurrentChat(1);
-        List<MessageImpl> history = generateHistoryFromCurrentChat(currentChat);
-        MessageHolder holder = newMessageHolderWithHistory(new History(history));
-        MessageListener messageListener = mock(MessageListener.class);
-        MessageTracker tracker = holder.newMessageTracker(messageListener);
-        MessageTracker.GetMessagesCallback callback = mock(MessageTracker.GetMessagesCallback.class);
-        ChatItem chat = new ChatItem();
-
-        holder.receiveHistoryUpdate(history, EMPTY_SET, EMPTY_RUNNABLE);
-
-        tracker.getNextMessages(1, callback);
-        verify(callback, times(1)).receive(history);
-        reset(callback);
-
-        holder.onChatReceive(null, chat, currentChat);
-        verify(messageListener, never()).messageChanged(any(Message.class), any(Message.class));
-        reset(messageListener);
-
-        holder.onMessageDeleted(currentChat.get(0).getIdInCurrentChat());
-        verify(messageListener, times(1)).messageRemoved(any(Message.class));
-        reset(messageListener);
-
-        holder.onChatReceive(chat, chat, currentChat);
-        verify(messageListener, times(1)).messageAdded(null, currentChat.get(0));
+        MessageImpl messageToDelete = currentMessages.get(5);
+        // Delete message from middle of chat
+        messageHolder.onMessageDeleted(messageToDelete.getServerSideId());
+        // Verify message deleted from middle
+        verify(messageListener, times(1)).messageRemoved(messageToDelete);
     }
 }
